+++ CHANGELOG.md
@@ -83,6 +83,11 @@ and this project adheres to
   The main effect of this is, when post-styling a table with CSS, cell content
   remains visible. I.e. Z-ordering is more intuitive.
 
+### Fixed
+
+- `dot` no longer drops regular edge labels from packed disconnected graphs.
+  #1616
+
 ## [15.1.0] – 2026-06-17
 
 ### Added